requirements.txt
# Data Generation
google-generativeai>=0.3.0
aiohttp>=3.9.0
tqdm>=4.66.0
Pillow>=10.0.0

# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0

# Transformers & VLM (Qwen2.5-VL Support)
# Note: Qwen2.5-VL requires transformers>=4.48.0 or install from source
transformers>=4.48.0
# OR install from source for latest Qwen2.5-VL support:
# git+https://github.com/huggingface/transformers.git
qwen-vl-utils>=0.0.8
accelerate>=0.33.0
peft>=0.13.0
bitsandbytes>=0.44.0

# Reinforcement Learning
trl>=0.12.0

# Training Utilities
datasets>=2.16.0
deepspeed>=0.15.0
einops>=0.8.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Logging & Monitoring
tensorboard>=2.15.0
wandb>=0.18.0

# Utilities
pyyaml>=6.0.1
numpy>=1.24.0
openai>=1.0.0

# Habitat Sim (optional - for rendering and data generation)
# Uncomment if you need habitat simulation:
# habitat-sim>=0.3.0
# habitat-lab>=0.3.0

# Development Tools (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
